Query tweaks
